Refactoring: move Keyword to parent package
