+++ java/code/src/com/redhat/rhn/frontend/action/channel/manage/PublishErrataHelper.java
@@ -23,7 +23,7 @@ import com.redhat.rhn.domain.channel.ChannelFactory;
 import com.redhat.rhn.domain.errata.Errata;
 import com.redhat.rhn.domain.errata.ErrataFactory;
 import com.redhat.rhn.domain.errata.Bug;
-import com.redhat.rhn.domain.errata.impl.Keyword;
+import com.redhat.rhn.domain.errata.Keyword;
 import com.redhat.rhn.domain.errata.ClonedErrata;
 import com.redhat.rhn.domain.org.Org;
 import com.redhat.rhn.domain.user.User;
