+++ java/code/src/com/redhat/rhn/frontend/action/errata/EditAction.java
@@ -24,7 +24,7 @@ import com.redhat.rhn.common.util.StringUtil;
 import com.redhat.rhn.domain.errata.Errata;
 import com.redhat.rhn.domain.errata.ErrataFactory;
 import com.redhat.rhn.domain.errata.Severity;
-import com.redhat.rhn.domain.errata.impl.Keyword;
+import com.redhat.rhn.domain.errata.Keyword;
 import com.redhat.rhn.domain.user.User;
 import com.redhat.rhn.frontend.action.common.BadParameterException;
 import com.redhat.rhn.frontend.struts.RequestContext;
