+++ java/code/src/com/redhat/rhn/frontend/xmlrpc/errata/ErrataHandler.java
@@ -40,7 +40,7 @@ import com.redhat.rhn.domain.errata.Errata;
 import com.redhat.rhn.domain.errata.ErrataFactory;
 import com.redhat.rhn.domain.errata.Severity;
 import com.redhat.rhn.domain.errata.Bug;
-import com.redhat.rhn.domain.errata.impl.Keyword;
+import com.redhat.rhn.domain.errata.Keyword;
 import com.redhat.rhn.domain.org.Org;
 import com.redhat.rhn.domain.rhnpackage.Package;
 import com.redhat.rhn.domain.rhnpackage.PackageFactory;
